Numerous unrelated changes
Chain of lambdas now also passes in the column names, rather than just
the column values, in `slurp.embelish`.

Added (average) relative momentum row via the above feature.  Here,
relative means relative to the rest of the portfolio, not all tickers
on the stock exchange or even the S&P500.  A small profile of stocks
would give meaningless values.  If downloading the S&P500 data is
cheap, then that would be a better litmus test to compare against.

Fields no longer require `getter` and `pullcast` (what typecasts the
data landing in that column) - this assumes that the data has already
been populated into the stocks dataframe in advance.

Migrated `trd0` and `momentum` to fields file.

The `sort_by` option now takes a list, and sorts by each in the order
supplied.

New "butterfly limit" option to the cli: a limit > 0 is as before, but
a limit < 0 will show the top-N and bottom-N rows.  Obviously this
only makes sense if the dataframe is also sorted based on something
meaningful.

The table printer function now manipulates data as a dataframe, then
passes it to `beautifultable` when everythin else is done (sorting,
filtering, and so on).