Add cache-update support and other changes
Rather than reuse the cache (fast, but stale), or regenearte the cache
(fresh, but slow), the cache can now be partially and incrementally
updated.  This feature is invoked anytime a query command contains a
specific list of stocks.

The stock-list parameter (-t) now supports a comma-separated list of
stocks too (rather than numerous `-t' parameters.

The `-C' (clear-cache) option has been removed.

The one AV endpoint used so far now supports the remotes throttling.
When the AlphaVantage API raises an exception, this client code will
no longer hit it for the remainder of the execution.  Additionally,
when tickers are supplied on the commandline, those specific stocks
are updated, otherwise, columns that are not set are prioritized over
others.

With the abovementioned change, the AV cache time has been reduced
from 30 days to 300 seconds.

Replace `collections.namedtuple' with `dataclasses.dataclass'.

Moved the measurement decorator to the debug module.

Deprecated the `S' or color-strip function, no longer required since
the presentation layer is no longer in the way of data parsing.

Helper function added to the repl.

Renamed ZERO to Z.

Minor bug fixes.