Move output filtering forward; consolidate fields
Filtering needs to take place earlier, otherwise it can lead to all
results being filtered out.  This was a bug; fixed by moving filtering
from step 6 to step 2.  Internally, this also means that filtering is
now done in Pandas, and not BeautifulTable.

With the ability to have multiple configurations per view, it is no
longer necessary to have so many views.  These have been consolidated.

New exception pattern to follow; add to and reraise the caught
exception, instead of masking it with a new exception.

Shorten huge numbers with 1000-suffixes, starting from a million, up
to and including a trillion.