Differentiate between a `you' and a `stock' score
The `you' score shows how well the investor has done with a stock.  It
may be a weak stock, but if bought cheap enough, if traded well in the
options market, and so on, it can still have a high score.

Conversely, a strong stock, bought at it's peak, may be a bad
investment.

Other general smaller changes.